get rid of interfaces package
