+++ pkg/hubbub/item.go
@@ -16,7 +16,6 @@ package hubbub
 
 import (
 	"fmt"
-	"github.com/google/triage-party/pkg/interfaces"
 	"github.com/google/triage-party/pkg/models"
 	"regexp"
 	"strconv"
@@ -44,7 +43,7 @@ var (
 )
 
 // createConversation creates a conversation from an issue-like
-func (h *Engine) createConversation(i interfaces.IItem, cs []*models.Comment, age time.Time) *Conversation {
+func (h *Engine) createConversation(i models.IItem, cs []*models.Comment, age time.Time) *Conversation {
 
 	authorIsMember := false
 	if h.isMember(i.GetUser().GetLogin(), i.GetAuthorAssociation()) {
