+++ pkg/hubbub/update_tracker.go
@@ -16,7 +16,7 @@ package hubbub
 
 import (
 	"fmt"
-	"github.com/google/triage-party/pkg/interfaces"
+	"github.com/google/triage-party/pkg/models"
 	"strings"
 	"time"
 
@@ -24,7 +24,7 @@ import (
 )
 
 // mtime is a workaround the GitHub misfeature that UpdatedAt is not incremented for cross-reference events
-func (h *Engine) mtime(i interfaces.IItem) time.Time {
+func (h *Engine) mtime(i models.IItem) time.Time {
 	return h.mtimeKey(i.GetUpdatedAt(), updateKey(i))
 }
 
@@ -53,7 +53,7 @@ func (h *Engine) mtimeKey(idea time.Time, key string) time.Time {
 	return updatedAt
 }
 
-func updateKey(i interfaces.IItem) string {
+func updateKey(i models.IItem) string {
 	// https://github.com/kubernetes/minikube/pull/8431
 	parts := strings.Split(i.GetHTMLURL(), "/")
 	if len(parts) < 7 {
@@ -67,7 +67,7 @@ func updateKey(i interfaces.IItem) string {
 	return fmt.Sprintf("%s/%s#%s", org, project, num)
 }
 
-func (h *Engine) updateMtime(i interfaces.IItem, t time.Time) {
+func (h *Engine) updateMtime(i models.IItem, t time.Time) {
 	key := updateKey(i)
 	h.updateMtimeByKey(key, t)
 }
