+++ pkg/models/item.go
@@ -1,27 +1,24 @@
-package interfaces
+package models
 
-import (
-	"github.com/google/triage-party/pkg/models"
-	"time"
-)
+import "time"
 
 // Item is an interface that matches both Issues and PullRequests
 type IItem interface {
 	// github specific
-	GetAssignee() *models.User
+	GetAssignee() *User
 	GetAuthorAssociation() string
 	GetBody() string
 	GetComments() int
 	GetHTMLURL() string
 	GetCreatedAt() time.Time
 	GetID() int64
-	GetMilestone() *models.Milestone
+	GetMilestone() *Milestone
 	GetNumber() int
 	GetClosedAt() time.Time
 	GetState() string
 	GetTitle() string
 	GetURL() string
 	GetUpdatedAt() time.Time
-	GetUser() *models.User
+	GetUser() *User
 	String() string
 }
